pyproject.toml
[tool.poetry]
name = "rehook"
version = "0.2.2"
description = ""
authors = ["Tom Punihaole <tom.punihaole@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "*"
djangorestframework = "*"
psycopg2-binary = "*"
django-filter = "*"
djangorestframework-xml = "*"
awsebcli = "^3.19.4"

[tool.poetry.dev-dependencies]
awsebcli = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
